GATK individual variant calling per-plate
Script to, on a single plate basis, create .g.vcf files for each individual, and then merge into a combined .vcf plate

We are doing this on a plate-by-plate basis because this step (especially the individual .vcf calculations) takes a long time. This way we can parallelize the process but submitted multiple jobs, one per plate.